Webscraped versions of Giant Eagle and ALDI's.
Added webscrapers for Giant Eagle and ALDI's for a specific location. They take two different approaches. The ALDI's webscraper simply finds the search box on the page, and searches the first 5 results to see if any of them may match with the ingredient we are looking for. The Giant Eagle one first webscrapes the entire site, and then any searches afterwards can parse through this webscraped result instead. This has the supposed advantage of speed versus storage (not tested yet, but should be true).